Persist redux, add animations & alerts

+++ web-build/component/FavoriteComponent.tsx
@@ -6,6 +6,7 @@ import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
 import Swipeout from 'react-native-swipeout';
 import { deleteFavorite } from '../redux/ActionCreators';
+import * as Animatable from 'react-native-animatable';
 
 const mapDispatchToProps = dispatch => ({
     deleteFavorite: (dishId) => dispatch(deleteFavorite(dishId))
@@ -30,19 +31,39 @@ class Favorites extends Component<any> {
                 {
                     text: 'Delete', 
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                        
+                    }
                 }
             ];
     
             return (
                 <Swipeout right={rightButton} autoClose={true}>
-                    <ListItem
-                        key={index}
-                        title={item.name}
-                        subtitle={item.description}
-                        onPress={() => navigation.navigate('DishDetail', { dishId: item.id })}
-                        leftAvatar={{ source: {uri: baseUrl + item.image}}}
-                        />
+                    <Animatable.View animation="fadeInRightBig" duration={2000}> 
+                        <ListItem
+                            key={index}
+                            title={item.name}
+                            subtitle={item.description}
+                            onPress={() => navigation.navigate('DishDetail', { dishId: item.id })}
+                            leftAvatar={{ source: {uri: baseUrl + item.image}}}
+                            />
+                    </Animatable.View>
                  </Swipeout>
             );
         };
